Update Main.java
Added a helper factory method to help create labels for the toppings. This should significantly cut down on code duplication. I mean look at how much code I was able to comment out!

I left pepperoni in the original format though since it's clear you still have plans for it. At least based on this listener you added to it. Eventually, this functionality will be added to the helper method, but I need you to define it before I can implement it.