Added a placeholder for inventory display
This is a stripped down version of the Ordering class just to demonstrate the basics. The idea is to provide a template for the inventory menu that anyone can work on as they see fit.

Right now it only loads and displays toppings from the database, but hopefully this should be enough to ignite people's imagination. It should be relatively easy to give the existing buttons proper functionality as well as to add more buttons.

Once that's done, it should be possible to copy the code in a way that allows you to load the other item types like sauces.