Added tabs to choose database item to display
Only toppings is present by default. Uncomment the buttons section to see all of the other tabs.

This is certainly a messy implementation.There's probably a more elegant implementation with for loops and polymorphism, but we're trying to hit a deadline. There shouldn't be too many changes to that section of the code thankfully though.

Main features of note: 
Click a button to run a function to display the relevant info in a panel specifically set aside for that purpose.
Click a different button to erase the contents of that panel, then populate it with the contents of the new category (if there are any).

There is a separate function for each category, just awaiting an implementation. There is an implementation for toppings, but that is still a WIP. These functions are very flexible. Separating it this way allows people to work on multiple implementations at a time.